Fix a typo in method name computation
It turns out we forget a "return" keyword when we want to return the
name of a method.  Ooops.

Fixed thus.

Please note that this patch doesn't come with its necessary update to
the tests/data/test-read-dwarf/test9-pr18818-clang.so.abi and
tests/data/test-read-dwarf/test22-pr19097-libstdc++.so.6.0.17.so.abi
files.  Those updates to regression tests are coming in a subsequent
patch.

	* src/abg-ir.cc (get_method_type_name): Really return the method
	name.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>