[dwarf-reader] Don't early canonicalize function types
Make sure we don't early-canonicalize function types.  We need to wait
after the end of building the ABI corpus before canonicalizing these.

Normally this should be taken care of by using
type_has_non_canonicalized_subtype but there are cases where that
function does not detect that a type has sub-types that are not
canonicalized.

Fixed thus.

	* src/abg-dwarf-reader.cc (maybe_canonicalize_type):
	Late-canonicalize function types.

signed-off-by: Dodji Seketeli <dodji@redhat.com>