[abixml writer] Make sure all function types are emitted
The abixml writer forgets to emit some function types that are only
referenced by other types.  Fixed thus.

Note that the necessary adjustment to reference outputs of abidw in
the regression test suite is not provided with this patch to ease
backporting.  Those adjustments are in a patch that comes at the end
of this patch series.

	* src/abg-writer.cc (write_translation_unit): Fix logic to avoid
	forgetting referenced function types.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>