Avoid emitting duplicated decls in abixml
This patch avoids emitting the same decl twice in the same ABI corpus
or corpus_group.  For the purpose of this duplication detection, a
decl is designated by its pretty representation.

The patch helps to reduce the size of the abixml files drastically by
cutting some of them, especially those representing the KMI of the
linux kernel in more than half.

	* src/abg-writer.cc (write_context::m_emitted_decls_map): New data
	member.
	(write_context::{decl_name_is_emitted, record_decl_as_emitted}):
	Define new memeber functions.
	(write_translation_unit): Do not emit a decl that has already been
	emitted.
	(write_var_decl, write_function): Record the decl as emitted.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>