Speedup comparison of decl-only classes
Profiling shows that while writting out the abixml for a big
corpus_group, comparing decl-only classes is a hot spot.

This patch avoids calling the function
class_or_union::priv_->comparison_started (which appears to be
culprit) in that case.

This makes writting out the abixml of the corpus_group of the Linux
kernel be ~20% faster.

	* src/abg-ir.cc (equals): In the overload for class_decl, if we
	are looking at a decl-only class, then directly call the equals
	function for class_or_union.  That one knows how to perform the
	comparison without calling the
	class_or_union::priv_->comparison_started function, in that case.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>