Fix some random deadlock while running fedabipkgidiff in tests
We are seeing some random cases where the regression test
runtestfedabipkgdiff.py hangs in what seems to be a deadlock.  This
seems to happen in fedabipkgidiff when it spawns a process to run
abipkgdiff.  More precisely, proc.communicate() hangs.

The documentation of that function at
https://docs.python.org/2/library/subprocess.html#subprocess.Popen.communicate
says:

    Note: The data read is buffered in memory, so do not use this
    method if the data size is large or unlimited.

So this patch avoids using proc.communicate() because the output of
abipkgdiff *can* be large.  Rather, the patch manually waits for the
the spawned abipkgdiff to finish, and then, read its output.

	* tools/fedabipkgdiff (abipkgidff): Do not use Popen.communicate()
	as it might hang if the data is large.  Rather, busy wait for the
	abipkgdiff process to finish and then get its output.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>