Ensure build_qualified_type can return non-qualified types
Const references are a useless redundancy because a reference is
always const.

Thus build_qualified_type can return a non-qualified type when it
figures out the qualifier is useless.

This patch adjusts build_qualified_type to recognize that.  It makes
it return a type_base_sptr rather than a qualified_type_def_sptr.

In build_ir_node_from_die, in the case where we handle qualified
types, the call to build_qualified_type is adjusted accordingly.

	* src/abg-dwarf-reader.cc (build_qualified_type): Return a
	type_base_sptr.
	(build_ir_node_from_die): Adjust the call to build_qualified_type.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>